Crate project_model

In rust-analyzer, we maintain a strict separation between pure abstract semantic project model and a concrete model of a particular build system.

Pure model is represented by the [base_db::CrateGraph] from another crate.

In this crate, we are concerned with “real world” project models.

Specifically, here we have a representation for a Cargo project (CargoWorkspace) and for manually specified layout (ProjectJson).

Roughly, the things we do here are:

Project discovery (where’s the relevant Cargo.toml for the current dir).

Custom build steps (build.rs code generation and compilation of procedural macros).

Lowering of concrete model to a [base_db::CrateGraph]

Modules

target_data_layout

Runs rustc --print target-spec-json to get the target_data_layout.

Structs

CargoConfig

CargoWorkspace

CargoWorkspace represents the logical structure of, well, a Cargo workspace. It pretty closely mirrors cargo metadata output.

ManifestPath

More or less [AbsPathBuf] with non-None parent.

PackageData

Information associated with a cargo crate

PackageDependency

PackageRoot

PackageRoot describes a package root folder. Which may be an external dependency, or a member of the current workspace.

ProjectJson

Roots and crates that compose this Rust project.

ProjectJsonData

Sysroot

TargetData

Information associated with a package’s target

WorkspaceBuildScripts

Enums

CargoFeatures

CfgOverrides

A set of cfg-overrides per crate.

InvocationLocation

InvocationStrategy

ProjectManifest

ProjectWorkspace

RustcSource

Describes how to set the rustc source directory.

TargetKind

UnsetTestCrates

Crates to disable #[cfg(test)] on.

Type Definitions

Package

Target