project_model

Module build_dependencies

Logic to invoke cargo for building build-dependencies (build scripts and proc-macros) as well as executing the build scripts to fetch required dependency information (OUT_DIR env var, extra cfg flags, etc).

In essence this just invokes cargo with the appropriate output format which we consume, but if enabled we will also use RUSTC_WRAPPER to only compile the build scripts and proc-macros and skip everything else.

Structs

BuildScriptOutput 🔒

Output of the build script and proc-macro building step for a concrete package.

WorkspaceBuildScripts

Output of the build script and proc-macro building steps for a workspace.

Enums

ProcMacroDylibPath

Functions

is_dylib 🔒