project_model

Struct CargoConfig

pub struct CargoConfig {

    pub all_targets: bool,
    pub features: CargoFeatures,
    pub target: Option<String>,
    pub sysroot: Option<RustLibSource>,
    pub sysroot_src: Option<AbsPathBuf>,
    pub rustc_source: Option<RustLibSource>,
    pub extra_includes: Vec<AbsPathBuf>,
    pub cfg_overrides: CfgOverrides,
    pub wrap_rustc_in_build_scripts: bool,
    pub run_build_script_command: Option<Vec<String>>,
    pub extra_args: Vec<String>,
    pub extra_env: FxHashMap<String, Option<String>>,
    pub invocation_strategy: InvocationStrategy,
    pub target_dir_config: TargetDirectoryConfig,
    pub set_test: bool,
    pub no_deps: bool,

}

Fields
all_targets: bool

Whether to pass --all-targets to cargo invocations.

features: CargoFeatures

List of features to activate.

target: Option<String>

rustc target

sysroot: Option<RustLibSource>

Sysroot loading behavior

sysroot_src: Option<AbsPathBuf>rustc_source: Option<RustLibSource>

rustc private crate source

extra_includes: Vec<AbsPathBuf>

Extra includes to add to the VFS.

cfg_overrides: CfgOverrideswrap_rustc_in_build_scripts: bool

Invoke cargo check through the RUSTC_WRAPPER.

run_build_script_command: Option<Vec<String>>

The command to run instead of cargo check for building build scripts.

extra_args: Vec<String>

Extra args to pass to the cargo command.

extra_env: FxHashMap<String, Option<String>>

Extra env vars to set when invoking the cargo command

invocation_strategy: InvocationStrategytarget_dir_config: TargetDirectoryConfig

Optional path to use instead of target when building

set_test: bool

Gate #[test] behind #[cfg(test)]

no_deps: bool

Load the project without any dependencies