project_model

Struct ManifestPath

pub struct ManifestPath { /* private fields */ }

More or less [AbsPathBuf] with non-None parent.

We use it to store path to Cargo.toml, as we frequently use the parent dir as a working directory to spawn various commands, and its nice to not have to .unwrap() everywhere.

This could have been named AbsNonRootPathBuf, as we donโt enforce that this stores manifest files in particular, but we only use this for manifests at the moment in practice.

Implementations

impl ManifestPath

pub fn parent(&self) -> &AbsPath

pub fn canonicalize(&self) -> !

pub fn is_rust_manifest(&self) -> bool