Struct project_model::ProjectWorkspace

pub struct ProjectWorkspace {
    pub kind: ProjectWorkspaceKind,
    pub sysroot: Sysroot,
    pub rustc_cfg: Vec<CfgAtom>,
    pub toolchain: Option<Version>,
    pub target_layout: TargetLayoutLoadResult,
    pub cfg_overrides: CfgOverrides,
}

Fields
kind: ProjectWorkspaceKindsysroot: Sysroot

The sysroot loaded for this workspace.

rustc_cfg: Vec<CfgAtom>

Holds cfg flags for the current target. We get those by running rustc --print cfg.

toolchain: Option<Version>

The toolchain version used by this workspace.

target_layout: TargetLayoutLoadResult

The target data layout queried for workspace.

cfg_overrides: CfgOverrides

A set of cfg overrides for this workspace.

Implementations

impl ProjectWorkspace

pub fn load( manifest: ProjectManifest, config: &CargoConfig, progress: &dyn Fn(String), ) -> Result<ProjectWorkspace>

pub fn load_inline( project_json: ProjectJson, target: Option<&str>, extra_env: &FxHashMap<String, String>, cfg_overrides: &CfgOverrides, ) -> ProjectWorkspace

pub fn load_detached_file( detached_file: &ManifestPath, config: &CargoConfig, ) -> Result<ProjectWorkspace>

pub fn load_detached_files( detached_files: Vec<ManifestPath>, config: &CargoConfig, ) -> Vec<Result<ProjectWorkspace>>

pub fn run_build_scripts( &self, config: &CargoConfig, progress: &dyn Fn(String), ) -> Result<WorkspaceBuildScripts>

Runs the build scripts for this ProjectWorkspace.

pub fn run_all_build_scripts( workspaces: &[ProjectWorkspace], config: &CargoConfig, progress: &dyn Fn(String), working_directory: &AbsPathBuf, ) -> Vec<Result<WorkspaceBuildScripts>>

Runs the build scripts for the given ProjectWorkspaces. Depending on the invocation strategy this may run a single build process for all project workspaces.

pub fn set_build_scripts(&mut self, bs: WorkspaceBuildScripts)

pub fn manifest_or_root(&self) -> &AbsPath

pub fn workspace_root(&self) -> &AbsPath

pub fn manifest(&self) -> Option<&ManifestPath>

pub fn buildfiles(&self) -> Vec<AbsPathBuf>

pub fn find_sysroot_proc_macro_srv(&self) -> Result<AbsPathBuf>

pub fn to_roots(&self) -> Vec<PackageRoot>

Returns the roots for the current ProjectWorkspace The return type contains the path and whether or not the root is a member of the current workspace

pub fn n_packages(&self) -> usize

pub fn to_crate_graph( &self, load: FileLoader<'_>, extra_env: &FxHashMap<String, String>, ) -> (CrateGraph, ProcMacroPaths)

pub fn eq_ignore_build_data(&self, other: &Self) -> bool

pub fn is_json(&self) -> bool

Returns true if the project workspace is Json.