project_model

Module sysroot

Loads “sysroot” crate.

One confusing point here is that normally sysroot is a bunch of .rlibs, but we can’t process .rlib and need source code instead. The source code is typically installed with rustup component add rust-src command.

Modules

stitched 🔒

Structs

Sysroot

Enums

RustLibSrcWorkspace

Functions

discover_rust_lib_src_dir 🔒

discover_rust_lib_src_dir_or_add_component 🔒

discover_sysroot_dir 🔒

get_rust_lib_src 🔒

get_rustc_src 🔒