Crate project_model

FIXME: write short doc here

Structs

BuildDataCollector

BuildDataResult

CargoConfig

CargoWorkspace
CargoWorkspace represents the logical structure of, well, a Cargo workspace. It pretty closely mirrors cargo metadata output.

PackageData
Information associated with a cargo crate

PackageDependency

PackageRoot
PackageRoot describes a package root folder. Which may be an external dependency, or a member of the current workspace.

ProcMacroClient

ProjectJson
Roots and crates that compose this Rust project.

ProjectJsonData

Sysroot

TargetData
Information associated with a package’s target

Enums

ProjectManifest

ProjectWorkspace

RustcSource
Describes how to set the rustc source directory.

TargetKind

Type Definitions

Package

Target