Enum project_model::ProjectWorkspace

pub enum ProjectWorkspace {
    Cargo {
        cargo: CargoWorkspace,
        build_scripts: WorkspaceBuildScripts,
        sysroot: Option<Sysroot>,
        rustc: Option<CargoWorkspace>,
        rustc_cfg: Vec<CfgFlag>,
        cfg_overrides: CfgOverrides,
        toolchain: Option<Version>,
    },
    Json {
        project: ProjectJson,
        sysroot: Option<Sysroot>,
        rustc_cfg: Vec<CfgFlag>,
    },
    DetachedFiles {
        files: Vec<AbsPathBuf>,
        sysroot: Sysroot,
        rustc_cfg: Vec<CfgFlag>,
    },
}

Variants

Cargo

Fields

cargo: CargoWorkspace

build_scripts: WorkspaceBuildScripts

sysroot: Option<Sysroot>

rustc: Option<CargoWorkspace>

rustc_cfg: Vec<CfgFlag>

Holds cfg flags for the current target. We get those by running rustc --print cfg.

FIXME: make this a per-crate map, as, eg, build.rs might have a different target.

cfg_overrides: CfgOverrides

toolchain: Option<Version>

Project workspace was discovered by running cargo metadata and rustc --print sysroot.

Json

Fields

project: ProjectJson

sysroot: Option<Sysroot>

rustc_cfg: Vec<CfgFlag>

Project workspace was manually specified using a rust-project.json file.

DetachedFiles

Fields

files: Vec<AbsPathBuf>

sysroot: Sysroot

rustc_cfg: Vec<CfgFlag>

Project with a set of disjoint files, not belonging to any particular workspace. Backed by basic sysroot crates for basic completion and highlighting.

Implementations

impl ProjectWorkspace

pub fn load(
manifest: ProjectManifest,
config: &CargoConfig,
progress: &dyn Fn(String)
) -> Result<ProjectWorkspace>

pub fn load_inline(
project_json: ProjectJson,
target: Option<&str>,
extra_env: &FxHashMap<String, String>
) -> Result<ProjectWorkspace>

pub fn load_detached_files(
detached_files: Vec<AbsPathBuf>
) -> Result<ProjectWorkspace>

pub fn run_build_scripts(
&self,
config: &CargoConfig,
progress: &dyn Fn(String)
) -> Result<WorkspaceBuildScripts>

Runs the build scripts for this ProjectWorkspace.

pub fn run_all_build_scripts(
workspaces: &[ProjectWorkspace],
config: &CargoConfig,
progress: &dyn Fn(String)
) -> Vec<Result<WorkspaceBuildScripts>>

Runs the build scripts for the given ProjectWorkspaces. Depending on the invocation strategy this may run a single build process for all project workspaces.

pub fn set_build_scripts(&mut self, bs: WorkspaceBuildScripts)

pub fn to_roots(&self) -> Vec<PackageRoot>

Returns the roots for the current ProjectWorkspace The return type contains the path and whether or not the root is a member of the current workspace

pub fn n_packages(&self) -> usize

pub fn to_crate_graph(
&self,
load_proc_macro: &mut dyn FnMut(&str, &AbsPath) -> ProcMacroLoadResult,
load: &mut dyn FnMut(&AbsPath) -> Option<FileId>,
extra_env: &FxHashMap<String, String>
) -> CrateGraph